+++ src/components/Layout/LayoutView.js
@@ -4,10 +4,9 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Header from '../Header/CustomizedHeader';
 import NavBar from '../NavBar/CustomizedNavBar';
 import Footer from '../Footer/CustomizedFooter';
-import Home from '../../pages/landing/landingView3';
+import Home from '../../pages/landing/landingView';
 import About from '../../pages/about/aboutView';
-// import News from '../../pages/news/newsView';
-import News from '../../pages/landing/landingView1';
+import News from '../../pages/news/newsView';
 import Ccdi from '../../pages/ccdi/ccdiView';
 import Error from '../../pages/error/Error';
 import Search from '../../pages/search/searchView';
